+++ contrib/oxl/osl/osl_canny_base.cxx
@@ -228,7 +228,7 @@ void osl_canny_base::Cluster_centre_of_gravity(int * const *jx, int * const *jy,
     //xc.reset(),yc.reset();xc.next(),yc.next();)
     //float newdist = hypot(x- *i/*xc.value()*/,y- *j/*yc.value()*/);
     double newdist;
-    { float dx = x- *i/*xc.value()*/, dy = y- *j/*yc.value()*/; newdist = vcl_sqrt(dx*dx + dy*dy); }
+    { double dx = x- *i/*xc.value()*/, dy = y- *j/*yc.value()*/; newdist = vcl_sqrt(dx*dx + dy*dy); }
     if ( dist<0 || newdist < dist ) {
       x0 = *i;//xc.value();
       y0 = *j;//yc.value();
